100% Test coverage
